requirements.txt
flask==2.3.3
flask-socketio==5.3.5
websocket-client==1.7.0
python-socketio[client]==5.10.0
python-dotenv==1.0.0
setuptools>=65.5.0